[build] Make building tests optional
Currently, the tests are built by default with the core library.  This is not
always desired behavior.  This changeset makes building the tests optional.

This is done by a new 'waf' command called 'test'.  Building of the tests is
disabled by default.  To build them type './waf test'.

Note, this does not always rebuild the whole tree.  For example, from a clean
tree the following:

  ./waf configure
  ./waf

will just build the core library and kernel modules.  If the above is followed
by (without an interleaving 'clean'):

  ./waf test

then just the test bits will be built.  Moreover, from a clean tree:

  ./waf configure
  ./waf test

will build the core library, kernel modules, and tests.

Signed-off-by: Meador Inge <meador_inge@mentor.com>
Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>