[shm] Add SysV-based SHM transport for easier development and debugging.
When the user specifies the --shm-sysv waf configure option, the SHM transport
will mmap() a file rather than physical RAM, and use SysV message queues rather
than IPIs. This code is architecture- and (theoretically) OS-agnostic; the
Linux kernel driver is not used.

The file size is currently hardcoded to 1MB, but this should be provided by
a more flexible configuration system in the near future.

Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>

---
If nobody comments sooner, I'll wait a couple days before committing this to
give folks time to review.

This transport works for the "hi" demo, but I'm not sure how reliable it is.
Also, there's some annoyance about the /tmp/mcapi-shm file persisting when
applications are killed in unusual ways...

In conjunction with the x86 atomic.h patch from last week, to build on an x86
host right now, one can:
	./waf configure --shm-sysv --no-kmods