[build] Consolidate and allow user to append CFLAGS.
Don't specify CFLAGS separately for libmcapi and libmcapi/shm.

Also, allow the user to easily provide debug options, e.g.
	CFLAGS='-g -O0' ./waf configure
which will be passed to gcc as "-O2 -Wall -g -O0".

Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>

---
The is obviously useful, but I'm not sure if this is the best way to accomplish
it with waf. Comments welcome.