[shm] Fix race in mcomm interrupt handler.
This patch fixes a race in the 'mcomm_interrupt' interrupt handler.  The race
can occur with the following sequence of events on a dual-core AMP system and
causes a particular interrupt to never be handled:

* core 1: mcomm_fd_ioctl_notify(..., 0)
* core 0: mcomm_interrupt
   * wake_up_interruptible(...)
   * returns IRQ_HANDLED
* core 0: mcomm_fs_ioctl_waite_read returns
* core 1: queues up another buffer
* core 0: shm_poll
   * pulls two buffers and drops the mbox count to 0
* core 1: mcomm_fd_ioctl_notify(..., 0)
* core 0: mcomm_interrupt
   * active == 0 (which implies mbox count == 0)
   * interrupt is *not* acked
   * returns IRQ_NONE
   * interrupts forever

This changeset fixes the race by always returning IRQ_HANDLED and by making
the IRQ unshared for mcomm interrupts.  Thanks to Hollis Blanchard for helping
me chase down the race and suggest a fix.

Signed-off-by: Meador Inge <meador_inge@mentor.com>
Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>