[test] Define and use an assert so the tests notify us of each failure.
Currently, the developer must manually go instrument each test, which is
unhelpful.

Also convert a few (immediately problematic) tests to use the new macro.

Also, stop counting errors. This isn't necessary (the developer can grep the
log output), and it's complicating the code surgery in this area. This
functionality can be reimplemented in the future if necessary.

Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>