[test] Invoke test cases directly, rather than spawn each in its own thread
We currently start a thread, then synchronously wait for it to finish before
spawning the next. Since there's no concurrent processing, we can just call
the test function directly.

Also pass the test function's return value back to the caller, so that we can report errors that way instead of through a global.

Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>