[libmcapi] Fix 'mcapi_tx_response' to use the right target port.
mcapi_encode_endpoint() is supposed to be passed a port number, not an
endpoint. As a consequence, 'mcapi_tx_response' was incorrectly sending
messages to the control RX port of the node *sending* the message, instead of
the node that is waiting for the message.

Ultimately, this caused 'mcapi_get_endpoint_i' to not work correctly when the
endpoint on the foreign node had not already been created.

Signed-off-by: Meador Inge <meador_inge@mentor.com>
Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>