[test] Revert changeset 7c9259f4fe08
It turns out that almost every caller of MCAPID_TX_Mgmt_Message() is using
inappropriate arguments, not just these two. Many of those callers don't
actually have an endpoint reference (so we couldn't just apply this style of
patch 700 times, even if we wanted to).

We'll have to fix the problem with an API change.

Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>