[test] management service works with ports, not endpoints
Almost all callers of MCAPID_TX_Mgmt_Message() in the frontend pass in a port
number, but MCAPID_Mgmt_Service() in the backend was trying to use it as an
endpoint. This worked by accident when the backend ran on node 0.

MCAPID_Mgmt_Service() now treats the parameter is a port number, and will
obtain an endpoint reference as necessary.

Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>