[test] Make scalar server expected size specification more explicit.
The functional test scalar server in '.../test/func/backend.c' expects to be
notified of what type of scalar it should expect to receive.  This notification
was being accomplished with 'MCAPID_TX_Mgmt_Message'.  However, this was
confusing as the 3rd parameter to 'MCAPID_TX_Mgmt_Message' is a foreign
port and the scalar server never uses this information.

This changeset creates a new helper function for scalar type notification to
encapsulate the oddity of the unused port parameter.

Signed-off-by: Meador Inge <meador_inge@mentor.com>
Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>