[RFC] [mcapi] Drop lock before calling transport->finalize()
The transport-level finalize() function must ensure that it's killed any
transport-level threads before returning, so we can safely proceed to tear down
OS-level and generic state.

However, it's possible that immedately after mcapi_finalize() took the global
lock, some data arrived, and a transport thread is blocked waiting for that
lock. In that case, transport->finalize() will deadlock if it waits for that
thread to exit.

To handle this, we drop the global lock before calling the transport level,
then re-acquire afterwards.

Signed-off-by: Hollis Blanchard <hollis_blanchard@mentor.com>