Functional test fixes
This is to fix a race condition in the functional test suite. The functional
test is supposed to overflow the shared memory buffer in the frontend, to
simulate ENO_BUFFER error for testing purposes. However it failed to get
ENO_BUFFER error due to race condition created by the backend.

Signed-off-by: Tien Hock Loh <thloh@altera.com>